ci: Enhance security of Rekor monitor and release workflows
This commit enhances the security and reliability of the Rekor monitor and release workflows.

The `rekor-monitor` workflow has been updated to include `packages: read` permissions, which are necessary to resolve container digests using the GitHub API.

The `release` workflow now compiles the SLSA generator from source by setting `compile-generator: true`. This is a security best practice that ensures the integrity of the generator.